manage booking api called

+++ src/pages/MyBooking/MyBooking.js
@@ -1,4 +1,3 @@
-import axios from 'axios';
 import React, { useEffect, useState } from 'react';
 import useAuth from '../../hooks/useAuth';
 import BookedItem from '../BookedItem/BookedItem';
@@ -19,28 +18,32 @@ const MyBooking = () => {
     }, []);
 
     const handleRemove = id => {
-        const url = `https://travel-mama-server.herokuapp.com/mybooking/${id}`;
-        fetch(url, {
-            method: "DELETE"
-        })
-            .then(res => res.json())
-            .then(data => {
-                if (data.deletedCount) {
-                    alert("Delete");
-                    const remaining = bookings.filter(booking => booking._id !== id);
-                    setBookings(remaining);
-                }
+        const sure = window.confirm('are you sure?');
+        if (sure) {
+            const url = `https://travel-mama-server.herokuapp.com/mybooking/${id}`;
+            fetch(url, {
+                method: "DELETE"
             })
+                .then(res => res.json())
+                .then(data => {
+                    if (data.deletedCount) {
+                        alert("Successfully deleted");
+                        const remaining = bookings.filter(booking => booking._id !== id);
+                        setBookings(remaining);
+                    }
+                })
+        }
     }
+
 
     if (isLoading) {
         return <Spinner></Spinner>
     }
 
     return (
-        <div className=' common-bg h-screen  '>
-            <div className='App pt-10'>
-                <h2 className='text-4xl'>Your Bookings</h2>
+        <div className=' common-bg md:h-screen'>
+            <div className='App p-10'>
+                <h2 className='text-4xl font-bold text-tomato'>Your Bookings</h2>
                 <div className='md:w-1/2 md:mx-auto md:flex flex-wrap'>
                     {
                         bookings.map(booking => <BookedItem key={bookings._id} booking={booking} handleRemove={handleRemove} ></BookedItem>)
